+++ server/src/test/kotlin/suwayomi/tachidesk/graphql/ChapterCountPerMangaTest.kt
@@ -1,0 +1,42 @@
+package suwayomi.tachidesk.graphql
+
+/*
+ * Copyright (C) Contributors to the Suwayomi project
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+import org.jetbrains.exposed.v1.jdbc.transactions.transaction
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+import suwayomi.tachidesk.graphql.dataLoaders.chapterCountPerManga
+import suwayomi.tachidesk.manga.model.table.ChapterTable
+import suwayomi.tachidesk.manga.model.table.MangaTable
+import suwayomi.tachidesk.test.ApplicationTest
+import suwayomi.tachidesk.test.clearTables
+import suwayomi.tachidesk.test.createChapters
+import suwayomi.tachidesk.test.createLibraryManga
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class ChapterCountPerMangaTest : ApplicationTest() {
+    @Test
+    fun `counts each manga's chapters`() {
+        val a = createLibraryManga("A")
+        val b = createLibraryManga("B")
+        val c = createLibraryManga("C")
+        createChapters(a, 7, read = false)
+        createChapters(b, 3, read = true)
+
+        val counts = transaction { chapterCountPerManga(listOf(a, b, c)) }
+
+        assertEquals(mapOf(a to 7, b to 3), counts)
+    }
+
+    @AfterEach
+    internal fun tearDown() {
+        clearTables(ChapterTable, MangaTable)
+    }
+}